Quality Management in Agile Projects
During the development processes in agile methodology, the teams have the responsibility to build quality alongside developing the product. This is ensured by putting into place the best engineering practices to build the products with highest quality. This case study focuses on tools and action to ensure product quality in agile processes.