Add files via upload
During the development processes in agile methodology, the teams have the responsibility to build quality alongside developing the product. This is ensured by putting into place the best engineering practices to build the products with highest quality. This is unlike the traditional method of project management where the quality is addressed reactively rather than proactively by the product testing and resolution of the issues. 

The quality management in agile projects plays a vital role as it is important to incorporate highest quality along with the processes so that with frequent changes, the quality of the output is maintained and does not get compromised.

In this report, I addressed risk management in agile processes, FMEA and the tools to ensure quality of agile product management.